Refactor UndoRedo to more proper OO
